requirements.txt
langchain_groq
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchain_huggingface
langchain_chroma